#6 fixed annoying bug of all variables are null!
