+++ src/main/java/backend/rest/services/AbstractService.java
@@ -1,6 +1,9 @@
 package backend.rest.services;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.ws.rs.container.ContainerRequestContext;
 import javax.ws.rs.core.Context;
+import javax.ws.rs.core.Request;
 import javax.ws.rs.core.UriInfo;
 
 /**
@@ -11,4 +14,12 @@ public class AbstractService {
     @Context
     protected UriInfo uriInfo;
 
+    @Context
+    protected ContainerRequestContext requestContext;
+
+    @Context
+    protected Request request;
+
+    @Context
+    protected HttpServletRequest httpServletRequest;
 }
